pyproject.toml
[project]
name = "mergeguard"
version = "0.1.0"
description = "Cross-PR intelligence for the agentic coding era"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["code-review", "pull-request", "ai-code", "git", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "tree-sitter-language-pack>=0.13.0",
    "pygithub>=2.5.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.42.0"]
all = ["mergeguard[llm]"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9.0",
    "mypy>=1.14",
    "respx>=0.22",
]

[project.scripts]
mergeguard = "mergeguard.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mergeguard"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
